Move joda-time classes and dependency to dedicated module
