+++ jpopulator-core/src/main/resources/META-INF/services/io.github.benas.jpopulator.api.RandomizerRegistry
@@ -1,2 +1,1 @@
 io.github.benas.jpopulator.randomizers.registry.InternalRandomizerRegistry
-io.github.benas.jpopulator.randomizers.registry.JodaRandomizerRegistry
