+++ jpopulator-joda-time/pom.xml
@@ -19,6 +19,11 @@
         </dependency>
 
         <dependency>
+            <groupId>joda-time</groupId>
+            <artifactId>joda-time</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>io.github.benas</groupId>
             <artifactId>jpopulator-test</artifactId>
             <scope>test</scope>
