+++ jpopulator-joda-time/src/main/resources/META-INF/services/io.github.benas.jpopulator.api.RandomizerRegistry
@@ -1,0 +1,1 @@
+io.github.benas.jpopulator.randomizers.registry.JodaRandomizerRegistry
